requirements.txt
python-dotenv==0.21.0
supabase==0.7.1
Flask==2.2.2
supabase-py==0.0.2
firebase_admin==6.0.1
anyio==3.6.2
atomicwrites==1.4.1
attrs==22.1.0
autopep8==2.0.0
bleach==5.0.1
certifi==2022.9.24
chardet==4.0.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
dataclasses==0.6
deprecation==2.1.0
docutils==0.19
dotty-dict==1.3.1
Flask==2.2.2
gitdb==4.0.10
GitPython==3.1.29
gotrue==0.5.4
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==2.10
importlib-metadata==5.1.0
iniconfig==1.1.1
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.11.0
MarkupSafe==2.1.1
more-itertools==9.0.0
packaging==21.3
pkginfo==1.9.2
pluggy==1.0.0
postgrest-py==0.10.3
py==1.11.0
pycodestyle==2.10.0
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pytest==6.2.5
python-dateutil==2.8.2
python-gitlab==3.12.0
python-semantic-release==7.32.1
pywin32-ctypes==0.2.0
readme-renderer==37.3
realtime==0.0.5
realtime-py==0.1.3
requests==2.25.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
storage3==0.3.5
supabase==0.7.1
supabase-py==0.0.2
supafunc==0.2.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
twine==3.8.0
typing_extensions==4.4.0
urllib3==1.26.13
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
zipp==3.11.0
gunicorn==20.1.0
